[NaCl SDK] Enable linux host build for nacl_io and nacl_io_tests.
The process of doing this already found one serious bug
in lseek and will enable us to find many more via asan and
other sanitisers which run on the linux host envionment.

R=binji@chromium.org, binji

Review URL: https://codereview.chromium.org/73083005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@235753 0039d316-1c4b-4281-b951-d872f2087c98

+++ native_client_sdk/src/tests/nacl_io_test/mount_html5fs_test.cc
@@ -453,7 +453,7 @@ TEST_F(MountHtml5FsTest, GetDents) {
     EXPECT_EQ(sizeof(dirent) * num_dirents, bytes_read);
 
     std::multiset<std::string> dirnames;
-    for (int i = 0; i < num_dirents; ++i) {
+    for (size_t i = 0; i < num_dirents; ++i) {
       EXPECT_EQ(sizeof(dirent), dirents[i].d_off);
       EXPECT_EQ(sizeof(dirent), dirents[i].d_reclen);
       dirnames.insert(dirents[i].d_name);
@@ -477,7 +477,7 @@ TEST_F(MountHtml5FsTest, GetDents) {
     EXPECT_EQ(sizeof(dirent) * num_dirents, bytes_read);
 
     std::multiset<std::string> dirnames;
-    for (int i = 0; i < num_dirents; ++i) {
+    for (size_t i = 0; i < num_dirents; ++i) {
       EXPECT_EQ(sizeof(dirent), dirents[i].d_off);
       EXPECT_EQ(sizeof(dirent), dirents[i].d_reclen);
       dirnames.insert(dirents[i].d_name);
